Move the selector
The following moves the selector into a common location so that we can
reuse it in the API server.

The code is really sound, except it doesn't handle revisions yes, that
can come later.
