+++ cmd/juju/charmhub/download.go
@@ -6,32 +6,24 @@ package charmhub
 import (
 	"context"
 	"fmt"
-	"math"
 	"net/url"
 	"os"
-	"sort"
 	"strings"
 	"syscall"
 
 	"github.com/juju/charm/v8"
 	"github.com/juju/cmd"
-	"github.com/juju/collections/set"
 	"github.com/juju/errors"
 	"github.com/juju/gnuflag"
 	"github.com/juju/os/v2/series"
 
 	"github.com/juju/juju/charmhub"
+	"github.com/juju/juju/charmhub/selector"
 	"github.com/juju/juju/charmhub/transport"
 	jujucmd "github.com/juju/juju/cmd"
 	"github.com/juju/juju/cmd/modelcmd"
 	corecharm "github.com/juju/juju/core/charm"
 	"github.com/juju/juju/environs/config"
-)
-
-const (
-	// DefaultReleaseChannel is the channel we should look at if a user didn't
-	// specify a channel.
-	DefaultReleaseChannel = "latest/stable"
 )
 
 const (
@@ -208,77 +200,34 @@ func (c *downloadCommand) Run(cmdContext *cmd.Context) error {
 		return errors.Trace(err)
 	}
 
+	channel := c.channel
 	// Locate a release that we would expect to be default. In this case
 	// we want to fall back to latest/stable. We don't want to use the
 	// info.DefaultRelease here as that isn't actually the default release,
 	// but instead the last release and that's not what we want.
-	if c.channel == "" {
-		c.channel = DefaultReleaseChannel
+	if channel == "" {
+		channel = corecharm.DefaultChannelString
 	}
-	charmChannel, err := corecharm.ParseChannelNormalize(c.channel)
+	charmChannel, err := corecharm.ParseChannelNormalize(channel)
 	if err != nil {
 		return errors.Trace(err)
 	}
 
-	var (
-		filterFn  FilterInfoChannelMapFunc
-		selectAll = c.arch == ArchAll && c.series == SeriesAll
-	)
-	switch {
-	case selectAll:
-		// Select every channel for all architectures and all series.
-		filterFn = func(transport.InfoChannelMap) bool {
-			return true
-		}
-	case c.arch != ArchAll && c.series == SeriesAll:
-		filterFn = filterByArchitecture(c.arch)
-	case c.arch == ArchAll && c.series != SeriesAll:
-		filterFn = filterBySeries(c.series)
-	default:
-		filterFn = filterByArchitectureAndSeries(c.arch, c.series)
-	}
-
-	// To allow users to download from closed channels (same UX for snaps), we
-	// link all closed channels to their parent channel, if their parent channel
-	// is also closed, we move to the next.
-	// We do the linking here because we want to ensure that we only ever filter
-	// once ensuring we're always correct.
-	channelMap, err := linkClosedChannels(info.ChannelMap)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	channelMap = filterInfoChannelMap(channelMap, filterFn)
-	revisions, found := locateRevisionByChannel(c.sortInfoChannelMap(channelMap), charmChannel)
-	if !found {
-		if c.series != "" {
-			return errors.Errorf("%s %q not found for %q channel matching %q series", info.Type, c.charmOrBundle, c.channel, c.series)
-		}
-		return errors.Errorf("%s %q not found with in the channel %q", info.Type, c.charmOrBundle, c.channel)
-	}
-
-	if selectAll {
-		// Ensure we have just one architecture.
-		if arches := listArchitectures(revisions); len(arches) > 1 {
-			list := strings.Join(arches, ",")
-			return errors.Errorf("multiple architectures (%s) found, use --arch flag to specify the one to download", list)
-		}
-	}
-	if len(revisions) == 0 {
-		return errors.Errorf("%s %q not found with in the channel %q", info.Type, c.charmOrBundle, c.channel)
-	}
-
-	revision := revisions[0]
-	resourceURL, err := url.Parse(revision.Download.URL)
-	if err != nil {
-		return errors.Trace(err)
-	}
+	selector := selector.NewSelectorForDownload(c.orderedSeries)
+	resourceURL, origin, err := selector.Locate(info, corecharm.Origin{
+		Channel: &charmChannel,
+		Platform: corecharm.Platform{
+			Architecture: c.arch,
+			Series:       c.series,
+		},
+	})
 
 	path := c.archivePath
 	if c.archivePath == "" {
 		path = fmt.Sprintf("%s.%s", info.Name, info.Type)
 	}
 
-	cmdContext.Infof("Fetching %s %q using %q channel at revision %d", info.Type, info.Name, charmChannel, revision.Revision)
+	cmdContext.Infof("Fetching %s %q using %q channel at revision %d", info.Type, info.Name, charmChannel, origin.Revision)
 
 	if err := client.Download(ctx, resourceURL, path); err != nil {
 		return errors.Trace(err)
@@ -326,20 +275,6 @@ func (c *downloadCommand) getCharmHubURL() (string, error) {
 	return charmHubURL, nil
 }
 
-func (c *downloadCommand) sortInfoChannelMap(in []transport.InfoChannelMap) []transport.InfoChannelMap {
-	// Order the channelMap by the ordered supported controller series. That
-	// way we'll always find the newest one first (hopefully the most
-	// supported).
-	// Then attempt to find the revision by a channel.
-	channelMap := channelMapBySeries{
-		channelMap: in,
-		series:     c.orderedSeries,
-	}
-	sort.Sort(channelMap)
-
-	return channelMap.channelMap
-}
-
 // CharmHubClient defines a charmhub client, used for querying the charmhub
 // store.
 type CharmHubClient interface {
@@ -370,212 +305,4 @@ type stdoutFileSystem struct{}
 // it is truncated.
 func (stdoutFileSystem) Create(string) (*os.File, error) {
 	return os.NewFile(uintptr(syscall.Stdout), "/dev/stdout"), nil
-}
-
-func filterByArchitecture(arch string) FilterInfoChannelMapFunc {
-	return func(channelMap transport.InfoChannelMap) bool {
-		if arch == ArchAll {
-			return true
-		}
-
-		platformArch := channelMap.Channel.Platform.Architecture
-		return (platformArch == arch || platformArch == ArchAll) ||
-			isArchInPlatforms(channelMap.Revision.Platforms, arch) ||
-			isArchInPlatforms(channelMap.Revision.Platforms, ArchAll)
-	}
-}
-
-func filterBySeries(series string) FilterInfoChannelMapFunc {
-	return func(channelMap transport.InfoChannelMap) bool {
-		if series == SeriesAll {
-			return true
-		}
-
-		platformSeries := channelMap.Channel.Platform.Series
-		return (platformSeries == series || platformSeries == SeriesAll) ||
-			isSeriesInPlatforms(channelMap.Revision.Platforms, series) ||
-			isSeriesInPlatforms(channelMap.Revision.Platforms, SeriesAll)
-	}
-}
-
-func filterByArchitectureAndSeries(arch, series string) FilterInfoChannelMapFunc {
-	return func(channelMap transport.InfoChannelMap) bool {
-		return filterByArchitecture(arch)(channelMap) &&
-			filterBySeries(series)(channelMap)
-	}
-}
-
-// FilterInfoChannelMapFunc is a type alias for representing a filter function.
-type FilterInfoChannelMapFunc func(channelMap transport.InfoChannelMap) bool
-
-func filterInfoChannelMap(in []transport.InfoChannelMap, fn FilterInfoChannelMapFunc) []transport.InfoChannelMap {
-	var filtered []transport.InfoChannelMap
-	for _, channelMap := range in {
-		if !fn(channelMap) {
-			continue
-		}
-		filtered = append(filtered, channelMap)
-	}
-	return filtered
-}
-
-func locateRevisionByChannel(channelMaps []transport.InfoChannelMap, channel corecharm.Channel) ([]transport.InfoRevision, bool) {
-	var (
-		revisions []transport.InfoRevision
-		found     bool
-	)
-	for _, channelMap := range channelMaps {
-		if rev, ok := locateRevisionByChannelMap(channelMap, channel); ok {
-			revisions = append(revisions, rev)
-			found = true
-		}
-	}
-	return revisions, found
-}
-
-func listArchitectures(revisions []transport.InfoRevision) []string {
-	arches := set.NewStrings()
-	for _, rev := range revisions {
-		for _, platform := range rev.Platforms {
-			arches.Add(platform.Architecture)
-		}
-	}
-	return arches.SortedValues()
-}
-
-type channelMapIndex struct {
-	witnessed  bool
-	channelMap transport.InfoChannelMap
-}
-
-func linkClosedChannels(channelMaps []transport.InfoChannelMap) ([]transport.InfoChannelMap, error) {
-	witness := make(map[string][]channelMapIndex)
-	for _, channelMap := range channelMaps {
-		track := channelMap.Channel.Track
-		if witness[track] == nil {
-			witness[track] = make([]channelMapIndex, len(channelRisks))
-		}
-		index := riskIndex(channelMap)
-		if index < 0 {
-			return nil, errors.Errorf("invalid risk %q", channelMap.Channel.Risk)
-		}
-		witness[track][index] = channelMapIndex{
-			witnessed:  true,
-			channelMap: channelMap,
-		}
-	}
-
-	secondary := make(map[string][]transport.InfoChannelMap)
-	for track, channels := range witness {
-		secondary[track] = make([]transport.InfoChannelMap, len(channelRisks))
-
-		for i, channel := range channels {
-			if channel.witnessed {
-				secondary[track][i] = channel.channelMap
-				continue
-			}
-
-			k := i
-			for ; k >= 0; k-- {
-				if channels[k].witnessed {
-					break
-				}
-			}
-
-			if k == -1 {
-				continue
-			}
-
-			link := channels[k].channelMap
-			link.Channel.Risk = channelRisks[i]
-			secondary[track][i] = link
-		}
-	}
-
-	var result []transport.InfoChannelMap
-	for _, risks := range secondary {
-		for _, channel := range risks {
-			result = append(result, channel)
-		}
-	}
-	return result, nil
-}
-
-func riskIndex(ch transport.InfoChannelMap) int {
-	for i, risk := range channelRisks {
-		if risk == ch.Channel.Risk {
-			return i
-		}
-	}
-	return -1
-}
-
-func isSeriesInPlatforms(platforms []transport.Platform, series string) bool {
-	for _, platform := range platforms {
-		if platform.Series == series {
-			return true
-		}
-	}
-	return false
-}
-
-func isArchInPlatforms(platforms []transport.Platform, arch string) bool {
-	for _, platform := range platforms {
-		if platform.Architecture == arch {
-			return true
-		}
-	}
-	return false
-}
-
-func constructChannelFromTrackAndRisk(track, risk string) (corecharm.Channel, error) {
-	rawChannel := fmt.Sprintf("%s/%s", track, risk)
-	if strings.HasPrefix(rawChannel, "/") {
-		rawChannel = rawChannel[1:]
-	} else if strings.HasSuffix(rawChannel, "/") {
-		rawChannel = rawChannel[:len(rawChannel)-1]
-	}
-	return corecharm.ParseChannelNormalize(rawChannel)
-}
-
-func locateRevisionByChannelMap(channelMap transport.InfoChannelMap, channel corecharm.Channel) (transport.InfoRevision, bool) {
-	charmChannel, err := constructChannelFromTrackAndRisk(channelMap.Channel.Track, channelMap.Channel.Risk)
-	if err != nil {
-		return transport.InfoRevision{}, false
-	}
-
-	// Check that we're an exact match.
-	if channel.Track == charmChannel.Track && channel.Risk == charmChannel.Risk {
-		return channelMap.Revision, true
-	}
-
-	return transport.InfoRevision{}, false
-}
-
-type channelMapBySeries struct {
-	channelMap []transport.InfoChannelMap
-	series     []string
-}
-
-func (s channelMapBySeries) Len() int {
-	return len(s.channelMap)
-}
-
-func (s channelMapBySeries) Swap(i, j int) {
-	s.channelMap[i], s.channelMap[j] = s.channelMap[j], s.channelMap[i]
-}
-
-func (s channelMapBySeries) Less(i, j int) bool {
-	idx1 := s.invertedIndexOf(s.channelMap[i].Channel.Platform.Series)
-	idx2 := s.invertedIndexOf(s.channelMap[j].Channel.Platform.Series)
-	return idx1 > idx2
-}
-
-func (s channelMapBySeries) invertedIndexOf(value string) int {
-	for k, i := range s.series {
-		if i == value {
-			return len(s.series) - k
-		}
-	}
-	return math.MinInt64
 }
